Put isIndividuValid and getEnfants in a separate common file
