+++ lib/simulation/openfisca/mapping/famille.js
@@ -1,5 +1,5 @@
 var _ = require('lodash');
-var mapping = require('./');
+var common = require('./common');
 
 module.exports = {
     parents: {
@@ -10,7 +10,7 @@ module.exports = {
         }
     },
     enfants: {
-        fn: mapping.getEnfants
+        fn: common.getEnfants
     },
     proprietaire_proche_famille: {
         fn: function(situation) {
