+++ lib/simulation/openfisca/mapping/foyerFiscal.js
@@ -1,6 +1,6 @@
 var _ = require('lodash');
 var moment = require('moment');
-var mapping = require('./');
+var common = require('./common');
 
 module.exports = {
     declarants: {
@@ -12,7 +12,7 @@ module.exports = {
     },
     // Aujourd'hui, dans mes-aides, tous les enfant et seuls les enfants sont transmis à Openfisca comme personnes à charge
     personnes_a_charge: {
-        fn: mapping.getEnfants
+        fn: common.getEnfants
     },
     rfr: {
         fn: function (situation) {
