+++ lib/simulation/openfisca/mapping/index.js
@@ -7,6 +7,7 @@ var ind = require('./individu');
 var men = require('./menage');
 var fam = require('./famille');
 var foy = require('./foyerFiscal');
+var common = require('./common');
 var PRESTATIONS = require('../prestations');
 
 var isNotValidValue = function(value) {
@@ -162,22 +163,9 @@ var transform = function(src, schema, situation) {
     return result;
 };
 
-var isIndividuValid = exports.isIndividuValid = function(individu, situation) {
-    var age = moment(situation.dateDeValeur).diff(moment(individu.dateDeNaissance), 'years');
-    var handicap = _.find(individu.situationsPro, { situation: 'handicap' });
-    return individu.role != 'enfant' || age <= 25 || handicap;
-};
-
-exports.getEnfants= function(situation) {
-    var enfants = _.filter(situation.individus, function(individu) {
-        return mapping.isIndividuValid(individu, situation) && individu.role == 'enfant';
-    });
-    return _.pluck(enfants, '_id');
-};
-
 var mapIndividus = exports.mapIndividus = function (situation) {
     var individus = _.filter(situation.individus, function(individu) {
-        return isIndividuValid(individu, situation);
+        return common.isIndividuValid(individu, situation);
     });
 
     return _.map(individus, function(individu) {
