+++ lib/simulation/openfisca/mapping/menage.js
@@ -1,5 +1,5 @@
 var _ = require('lodash');
-var mapping = require('./');
+var common = require('./common');
 
 module.exports = {
     personne_de_reference: {
@@ -14,7 +14,7 @@ module.exports = {
         }
     },
     enfants: {
-        fn: mapping.getEnfants
+        fn: common.getEnfants
     },
     statut_occupation: {
         fn: function(logement) {
